Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ayan Naskar <ayannaskar5067@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
spacy = "^3.8.4"
fastapi = "^0.115.11"
python-docx = "^1.1.2"
pdfplumber = "^0.11.5"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
